Successfully sorted T0Black
I separated "Well" into "PlateRow" (which holds a letter that signifies the row of the plate that well is in) and "PlateCol" (which holds a number that signifies the column of the plate that well is in. I also created a new column called "Plot" which holds the plot number. I then sorted T0Black using "PlateCol" and "Plot" using the sort_values() method.

I also created dataframes of the standards (quench control) and the homogenate control. This is only the beginning of the process for determining which values in the T0Black dataframe falls under which form of control. I will reference the German et al 2011 paper for this process. Successfully determining which values fall under which form of control will allow me to better wrangle the T0Black dataframe until I can start calculating enzyme activities.