Added a memo to describe how to deal with homogenate control values
After looking through Steve's script (to viewers who do not know, I am referring to a script that my PI wrote that I used as an example for the calculateEnzymeActivity.py script), I couldn't find any lines that use column 10 of the BUFFER plates for homogenate control. Instead, it seems that only column 10 of the SAMPLE plates are used for homogenate control. I wrote a memo that basically described that. So from now on, I will use column 10 of the sample plates for homogenate control.