Shortened how the dataframe containing standard fluorescence values are made
Instead of manually setting the "PlateCol" field of each of the 7 dataframes, I put them into a list and then looped through the list using a for loop. I then fed this list into pd.concat() so that all 7 dataframes can be concatenated simultaenously instead of writing multiple lines that only concatenate 2 dataframes at a time.