Write methods in module
The code that I used to wrangle T0Black, I'm writing them in the enzymeWrangling.py module to make wrangling enzyme data more straightforward. I can simply call the methods in this module, and they will return wrangled versions of the enzyme dataframe.

I intend that the methods in the module be called from top to bottom. The very top method must be called first, followed by the one below it, and so on, so forth.