Added research ideas
I wrote Hypotheses.md to describe how I'm framing this study and the hypotheses I will be testing. I also added a description of how oxidative enzyme assays are carried out in "Enzyme Assays Description.md".

In addition, I changed my conclusions in the markdown file discussing normalizing enzyme activity. I realized that the parameters that changed are enzyme concentrations, not k2 or Km. While normalized k2 and Km will surely be different, it's kinda weird to divide constants by litter mass. Constants are constants and should be left alone, so dividing them to normalizing them feels off. In addition, in the figure I used to analyze 3 hypothetical samples, are the differences due to differences in k2? Or differences in Km? Like if you compare the middle and right-hand column, what are the differences attributed to? Well, the differences, here, are actually due to enzyme concentrations. It just takes a different amount of litter to achieve the same enzyme concentration, and it's this that causes differences in normalized enzyme activity. k2 and Km are the same between all 3 samples. It just so happens that each sample takes a different amount of litter to achieve the same amount of enzymes.