Worked on calculating T5 activities
There were some data entry errors with this timepoint as well as samples that have been assayed at least twice. Details of this had been written in the calculateEnzymeActivity.py script. I finished calculating the oxidase activities of T5, and had finished calculating the hydrolase activities for all but 1 replicate of 8LXX for this timepoint.