Finish calculating T5 hydrolase activity
I found out that the first time that 8LXX was assayed for hydrolase activity, it had no data, hence why it was assayed twice. I just removed the parts with no data and kept the other time the sample was assayed in the main T5hydro dataframe. Then I was able to successfully calculate and plot T5 hydrolase activity.