Judging T5 enzyme data
I classified error types in the T5 data after looking at the graphs of enzyme activity. I judged the graphs to see if the data looks noisy or not, if there is any signs of substrate inhibition, or, in the case of peroxidase, if there is a decreasing trend in enzyme activity as substrate concentration increases.

Steve previously suggested that if the data looks especially noisy, then that that's a sign of low enzyme activity, so low that the plate reader only picked up noise. So I'm going with that suggestion for now. So, when I see noisy data, it generally makes me believe that there is low enzyme activity.