Tried fitting ECA but ran into model identifiability problems
A model's parameters are identifiable if a small change in a single parameter produces a noticeable change in the model's predicted values. The more identifiable the model, the easier it is to find the true values of the parameters.

I tried fitting the single-enzyme, single-substrate formulation of the equilibrium chemistry approximation (ECA) to alpha-glucosidase activity of sample 4LXX from time point T0 and I ran into the problem of the parameters not being identifiable. I managed to find 2 very different sets of parameters that gives almost identical predictions. Steve suggested that I use Michaelis-Menten instead of ECA, which is a real, real shame. But, fine.

I'm gonna need to rework my hypotheses. Let's see if I can keep the same hypotheses and change only the predictions, or I might need to come up with new hypotheses entirely.