Finished abstracting the nonlinear regression process
I finished (or at least that's what I like to believe) the process for abstracting the nonlinear regression of the Michaelis-Menten equation. The function should be applicable to both hydrolytic and oxidative enzymes. So, now, in order to get enzyme parameters for both hydrolytic and oxidative enzymes, I just need to clean the data for substrate inhibition and set negative activity values as 0 and then call this function.