Wrote function to plot regression
This function plots both the calculated activity values and also the regression for a sample's enzyme using the fitted Michaelis-Menten parameters for that sample's enzyme. I intend this function to be as general as possible, that I can perform multiple data cleaning steps to clean calculated enzyme activities and yet can still plot them using this function. I'm going to add in the ability to save figures later to this function.