Extensive modifications & debugging of nonlinRegress() and plotRegress()
These 2 functions, before today, would have suffered from problems concerning the value of "can't fit" in the MM parameters dataframes. The try except statements in the nonlinRegress() function would be problematic, as the try statement would have fitted Michaelis-Menten to both PER & PPO. If the fitting of any of these 2 enzymes fail, then the except statement runs, and we would not know which enzyme specifically did the fitting fail. So, to fix that, I separated attempts to fix PER and PPO into separate try except statements, with 1 statement for PER and 1 statement for PPO. Now, if we run the nonlinRegress function on an oxidase activity dataframe and there's a failure, then we should know which enzyme & which replicate failed.

I've also modified plotRegress() to better deal with samples containing enzymes that failed to fit Michaelis-Menten. I added the condition of checking the Vmax of every enzyme of every sample; if the Vmax of a sample's enzyme is "can't fit", then that sample's enzyme's actual & estimated activities won't be plotted.

plotRegress() is also extensively modified for the replicates of oxidase activity of sample 47RRX in T5. The way that I generated the dataframe of calculated oxidase activities for T5 is that the replicates are either 1 or 2. Thus, even though sample 47RRX was assayed twice for oxidase activity, both assays have replicates of 1 or 2 in this dataframe. However, the way that I intend to generate the dataframe of oxidase parameters for T5 is that I want to have replicates of 47RRX be in the range of [1, 4]. Thus, there is a difference between how replicates were recorded in the calculated oxidase activity dataframe of T5 and a difference in how I want replicates to be recorded in the oxidase parameters dataframe of T5. To account for this difference in replicates, I essentially created 2 replicate variables in dealing with sample 47RRX in the calculated T5 oxidase activity dataframe: 1 variable used for indexing the calculated activities dataframe and 1 variable used for indexing the dataframe of parameters.