Dealing with possible issues that arise from substrate inhibition in oxidase data
The way the dataframe of calculated oxidase activity is set up is that for each replicate of a sample, there will be 2 columns side by side, 1 column of PPO activity and 1 column of PER activity. As a result, I can't just throw away rows that show substrate inhibition like I can with hydrolase activity.

I wrote a function to clean oxidase data, and then I wrote a function that abstracts the cleaning process (at least phase 1 of the cleaning process, which is removal of negative activities and taking out substrate inhibition and then re-examining the data once these are taken out) so that a dataframe of either hydrolase or oxidase activity from a single time point can be cleaned by calling this function.

I also fixed up the nonlinear regression function and the function for plotting the parameters with the cleaned, actual activities, to take into account substrate inhibition of oxidase activity. I made it so that they would only perform nonlinear regression on and plot the cleaned data that does not have any substrate inhibition and negative enzyme activities.

I also changed the "Samples with errors.xlsx" file by changing one of the oxidase samples to include substrate inhibition, as on reexamination of that sample, I believe that one of its enzyme replicates show substrate inhibition.