Abandon acclimation testing
So I decided to stop testing for acclimation for a couple of reasons. First reason is, in order to test that, I would need to get in-situ enzyme activity. I thought about doing so using leaf litter chemistry, but after I thought about it, this seems impossible. These enzymes' substrates are polymers, and to calculate the substrate concentrations, you would need to get the amount of moles of these polymers. However, how are you supposed to get the amount of moles in the first place? These substrates are polymers here, and a polymer is a chain of monomers, so a mole of a polymer is 6.02 * 10^23 chains of monomers, and who's to say that these chains are all the same length? So, no. Second, we don't have direct in-situ activity data; our only possible option was to use litter chemistry data, which I have already said that we can't. Third is, I'm using Michaelis-Menten, not the Equlibrium Chemistry Approximation (ECA). If I'm using ECA, then I can look directly at k2, and Km is likely to have a different definition from how it's currently used in extracellular enzyme assays in microbial ecology, a more molecular definition that describes the actual enzyme kinetics. Both k2 and Km -- I don't know about Km, but I'm more sure about k2 -- in ECA describe enzyme kinetics as opposed to enzyme and product concentrations. However, since I'm using Michaelis-Menten, the way that these parameters have been traditionally used in microbial ecology is to describe enzyme and product concentrations.

So, I abandoned acclimation testing and I rephrased my first research question into how substrate/product concentrations will change following changes in enzyme concentrations. My hypothesis is that enzyme concentrations determine product concentrations in a way that they're correlated, and this leads to the same predictions as my initial version of Hypothesis 1, but this hypothesis is more in line with the literature. Otherwise, I don't have to change my other hypotheses.