Started testing Michaelis-Menten parameters for normality
I do this testing using the Shapiro-Wilk test -- the default null hypothesis of which is that a sample (using the statistics definition here, which is a group of observations subsetted from a larger population) is from a normally distributed population. The purpose of this test is to see which parameters need to be transformed to ensure that ANOVA can be used. So far, I've got the test results for the parameters when no data transformation had taken place yet, but I haven't exported these results into an Excel file yet.