Began performing three-way, factorial ANOVAs
I've managed to do so for Vmax of AG, and I found that neither of the treatments and time points by themselves influence AG Vmax and none of the interactions between these 3 independent variables influence AG Vmax. Time to repeat the process for AG Km and the remaining enzymes (7 other enzymes).