Tested some syntax in running ANOVAs
I was trying to see whether to use the "Sum" thing in the formulas for my ANOVAs, and whether to use the C() thing as well as the differences between asterisks and colons. It seems like if I'm doing a type 3 ANOVA, I would need to incorporate the "Sum" thing, and in order to use it, I need to use the "C()" thing as well. So, to do a type 3 ANOVA, I would need to use both "Sum" and "C()". If I'm only doing a type 2 ANOVA, then I only need to use the "C()" thing for a factor whose values are numerical when the factor is actually categorical and the numbers stand for discrete categories.

I also moved the AG Vmax ANOVA syntax below the MANOVA chunk.