Began Tukey posthoc testing
I performed a Tukey posthoc test for AP Vmax as a proof of concept for a function to abstract the process. Ran into a couple of problems with the function so far.

(1) Function does not distinguish between non-significant and significant interactions as I intended.

(2) Excel files that function writes to  only has 1 sheet for 1 parameter when some enzymes have significant interactions or main effects for both parameters, so these enzymes should have 2 sheets.

(3) index and and "Unnamed: 7" column are both printed in these spreadsheets when they should not be.