Made most of boxplots, part 1
Yeah so, I made most of the boxplots, yay. I also just realized that since some of the interactions or main effects that are significant by ANOVA but insignificant by Tukey, that frees up me to test certain main effects that I couldn't test before due to these main effects being part of one of those aforementioned interactions that are significant by ANOVA but insignificant by Tukey. So I will review my Tukey posthoc and ANOVA results again to see which of these main effects can I also test for.

For the boxplots that I made so far, I made some more modifications to them. I removed the outliers that are outside of the horizontal caps and vertical whiskers to compress the y-axis a bit. I also made the median lines thick and cyan so that they are more visible.

I'm gonna push these figures up a few at a time.