Revising figures
Steve's comments on my first draft said that my axis and tick labels are too small, so I changed the boxplots.py script a bit to make the fonts of both the axes and ticks larger. Turns out that for the three-way boxplot for PPO Vmax, I can't make the x-tick labels as large as for the other boxplots, so that's too bad. I also made the boxplots a bit shorter (in height) so that I can fit more boxplots on a page of my thesis.

I also reformatted the figures on my thesis Word doc a bit. Deleted any mentions of linear regression in my thesis, including the linear regression figure. I also put boxplots of both Vmax and Km of the same enzyme on the same page as part of the same figure, and I label each subplot with a letter in a text box. Because of these changes, I also had to change the numbering of my figures as well. I then made these changes in my results and discussion.