First meeting with Steve and Ashish to discuss writing a paper
We will write a paper that will focus on enzymes specifically. We will look at metagenomic data that tells us how many genes that encode for CAZymes are there in this project and compare them to enzyme activity data and, hopefully, mass loss data if Ashish sends them in. In addition, this paper will focus on individual substrates (such as hemicellulose, cellulose, chitin, lignin, etc).