Conducted Tukey post-hoc on litter chemistry data
Got raw Tukey post-hoc results. I spent a couple hours trying to create a compact letter display (just assigning letters to Tukey groups to designate treatments that are significantly different from each other) and it was a pain, so I found a script that can do the job. Gonna fix up the script and possibly apply the script to my Tukey results for Vmax and Km.