Edited cld.py module to make it look cleaner
As the title had said. the original script looks like shit, very hard to read. I edited it to make it look easier (at least for me) to read. I also renamed some of the columns referenced in the module to match the output of statsmodel's Tukey test results. And I removed a condition that checks if the format of the p-values are strings.