Check enzyme compact letter displays
As I originally created compact letter displays for enzymes by hand, I wrote a script to check if these compact letter displays match with the raw Tukey test results. And they do match, so I did not make any mistakes in creating these compact letter displays by hand in the past. So I don't need to make any updates to these on the shared drive with Steve and Ashish.

I also created a new main() function in cld.py that both creates a compact letter display and checks it to see if the compact letter display matches the raw Tukey results used to create it.