Test Box-Cox transformation
I was afraid that performing a Box Cox transformation over multiple treatment groups will transform the data so that they all belong to one population. If that is the case, then a multiple comparison would not be able to detect differences between any treatments. To see if this fear is justified, I tested the Box-Cox transformation on amide1 and glycosidic bond FTIR spectral area. The transformation did not result in a loss of differences between groups. As a result, I can safely use the Box-Cox transformation to transform non-normal data.