Linear mixed models for enzyme Vmax
After previously converting time to a continuous variable and re-running the models for some enzymes, I finished re-running the models for the enzymes that I have previously ran with time as a categorical variable. I also finished running the new model for all enzymes.

The model for PPO has a slight difference from the previous enzymes in that the assays for PPO has a replicate, meaning that each plot from a particular time point has its PPO Vmax measured twice (for plot 47RRX from time point 5, for some reason, it was measured 4 times). Therefore, I included the replicate (which is a categorical variable) as a variance component, nested under the overall group of plot ID.