Refractoring
I used a string.format() method to create a formula for a linear mixed effects model instead of writing a series of if-else statements.

I created a function to calculate Cohen's D for all the main effects for all enzymes within the Vmax dataset or all all functional groups within the litter chemistry dataset to abstract the process a bit.

I exported the linear mixed model effect results out.