Finished cleaning new metagenomic dataset
The original "id" column in the file contains information on a plot's vegetation, precipitation, and time point. The cleaning process involves checking if these 3 pieces of information from the original "id" column makes sense, and then the plot ID itself was obtained from this column. And these 3 pieces of information makes sense, so the data he handed me was internally consistent. I produced a .csv file of the cleaned up data and began analyzing it.