Finished re-analysis
Made slight changes to the function that updates the linear mixed effect model results. Previously the function only updates results for dependent variables that were found to be significant under Tukey's while ignoring results for dependent variables that were tested under Tukey's but were not found to be significant. Changed the function slightly so that it also updates dependent variables that have been tested but weren't found to be signifcant.

Added some notes to explain the additional Tukey's I performed for some dependent - independent variable combinations.

Created an Excel file of the re-analysis results. I'm now completely done with the re-analysis